241029/정윤석/table 출력 기능, table에 정보 입력 기능 추가

+++ ChessTrainingAI/Assets/Scripts/Class/Piece/Piece.cs
@@ -12,7 +12,7 @@ public abstract class Piece : MonoBehaviour
 {
     #region 기물 정보
     [Header("기물 정보")]
-    public int piecePoint;
+    public float piecePoint;
     public bool isMovable;
     public PieceType pieceType;
     public GameColor pieceColor;
